refactor : move common components inside common package
