+++ src/main/java/coz/weavon/common/exception/model/RestException.java
@@ -1,6 +1,6 @@
-package coz.weavon.exception.model;
+package coz.weavon.common.exception.model;
 
-import coz.weavon.message.Translator;
+import coz.weavon.common.message.Translator;
 import coz.weavon.provider.ApplicationContextProvider;
 
 public class RestException extends RuntimeException {
