+++ src/main/java/coz/weavon/common/io/validator/RangeValidator.java
@@ -1,6 +1,6 @@
-package coz.weavon.io.validator;
+package coz.weavon.common.io.validator;
 
-import coz.weavon.util.DateTimeUtils;
+import coz.weavon.common.util.DateTimeUtils;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import java.lang.reflect.Field;
